requirements.txt
anyio==3.6.2
asgiref==3.5.2
astroid==2.12.10
asttokens==2.2.1
autopep8==1.7.0
backcall==0.2.0
beautifulsoup4==4.11.1
Brlapi==0.8.2
Brotli==1.0.9
cefpython3==66.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.2
contourpy==1.0.7
cryptography==3.3.2
cupshelpers==1.0
customtkinter==5.0.3
cycler==0.11.0
darkdetect==0.8.0
dbus-python==1.2.16
debugpy==1.6.5
decorator==5.1.1
dill==0.3.5.1
distlib==0.3.6
distro-info==1.0
Django==4.0.2
django-ckeditor==6.2.0
django-js-asset==2.0.0
dnspython==2.0.0
email-validator==1.3.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.89.1
filelock==3.8.0
Flask==2.2.2
fonttools==4.38.0
fuse-python==1.0.2
gpg===1.14.0-unknown
gyp==0.1
h11==0.14.0
httpcore==0.16.3
httplib2==0.18.1
httptools==0.5.0
httpx==0.23.3
ibm-db==3.1.4
idna==2.10
importlib-metadata==4.12.0
ipykernel==6.19.4
ipython==8.8.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==7.4.8
jupyter_core==5.1.2
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
louis==3.16.0
lxml==4.9.1
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mccabe==0.7.0
MouseInfo==0.1.3
multitasking==0.0.11
mutagen==1.45.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.23.3
openpyxl==3.0.10
orjson==3.8.5
packaging==22.0
pandas==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipenv==2022.9.24
platformdirs==2.5.2
poo==1.0.0
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
PyAudio==0.2.11
PyAutoGUI==0.9.53
pycairo==1.16.2
pycodestyle==2.9.1
pycryptodomex==3.14.1
pycups==2.0.1
pycurl==7.43.0.6
pydantic==1.10.4
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.7.1
PyGObject==3.38.0
pyjokes==0.6.0
pylibacl==0.6.0
pylint==2.15.2
PyMsgBox==1.0.9
pyOpenSSL==20.0.1
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.2
PyQt5-sip==12.8.1
PyRect==0.2.0
PyScreeze==0.1.28
PySimpleSOAP==1.16.2
pysmbc==1.0.23
python-apt==2.2.1
python-dateutil==2.8.2
python-debian==0.1.39
python-debianbts==3.1.0
python-dotenv==0.21.0
python-multipart==0.0.5
python3-xlib==0.15
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.2.1
pywhatkit==5.4
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.3.1
pyzmq==24.0.1
reportbug==7.10.3+deb11u1
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.22.0
tk==0.1.0
tkinterhtml==0.7
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
tornado==6.1
traitlets==5.8.0
ttkbootstrap==1.10.0
ttkthemes==3.2.2
typing_extensions==4.3.0
ujson==5.7.0
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.16.5
virtualenv-clone==0.5.7
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
Werkzeug==2.2.2
wikipedia==1.4.0
wrapt==1.14.1
xdg==5
xlrd==2.0.1
yfinance==0.1.74
yt-dlp==2022.4.8
zipp==3.8.1
